frontend/requirements.txt
streamlit
requests
pillow
boto3
dotenv